backend/scrapers/crawl4ai_scrapers/requirements.txt
crawl4ai>=0.5.0
python-dotenv==1.0.1
aiohttp==3.9.1
aiofiles==23.2.1 